+++ tests/Assets/Loop-Manual.spec.js
